Changed variable naming of multiple choice item on android
Multiple choice variable naming will be moved to use item indices instead of text, to be language independent. This commit changes the item in the android application to save the additional variables this way. This code is conditional on the server version, in order to retain backwards compatibility with older servers.