fixed an object copy bug, preventing memory overflow
MainActivity on Android would retreive a separate questionnaire object from DbLogic for each page of a questionnaire, which prevented the caching mechanism in the Questionnaire class to work properly. This lead to each Questionnaire page reloading the complete JSON source.
In studies with larger source and a large amount of pages this could lead to a crash due to memory overflow.