added forced study update after db update
This is to avoid a possible edge case where a study using the new features is joined before the client update, thereby using default empty scriptEndBlock and virtualInputs. These parts of the json require an update, but ESMira will think the study is already up to date. A forced study update should prevent this in most cases.